Use rollup to build modules
