+++ .github/workflows/nodejs.yml
@@ -69,7 +69,7 @@ jobs:
       - name: lint
         run: yarn lint
 
-  build-modules:
+  build-rollup:
     runs-on: ubuntu-latest
     env:
       CI: true
@@ -97,8 +97,39 @@ jobs:
             ${{ runner.OS }}-node_modules-
       - name: yarn install
         run: yarn install
-      - name: build-modules
-        run: yarn build-modules
+      - name: build-rollup
+        run: yarn build-rollup
+
+  build-tsc:
+    runs-on: ubuntu-latest
+    env:
+      CI: true
+    steps:
+      - uses: actions/checkout@v1
+      - name: Use Node.js 16.x
+        uses: actions/setup-node@v1
+        with:
+          node-version: 16.x
+      - name: Get yarn cache
+        id: yarn-cache
+        run: echo "::set-output name=dir::$(yarn cache dir)"
+      - uses: actions/cache@v1
+        with:
+          path: ${{ steps.yarn-cache.outputs.dir }}
+          key: ${{ runner.os }}-yarn-${{ hashFiles('**/yarn.lock') }}
+          restore-keys: |
+            ${{ runner.os }}-yarn-
+      - name: Cache node modules
+        uses: actions/cache@v1
+        with:
+          path: node_modules
+          key: ${{ runner.OS }}-build-${{ hashFiles('**/yarn.lock') }}
+          restore-keys: |
+            ${{ runner.OS }}-node_modules-
+      - name: yarn install
+        run: yarn install
+      - name: build-tsc
+        run: yarn build-tsc
 
   test:
     runs-on: ubuntu-latest
