+++ package.json
@@ -15,6 +15,8 @@
     "@babel/preset-typescript": "^7.12.16",
     "@codechecks/client": "^0.1.10",
     "@hot-loader/react-dom": "^16.11.0",
+    "@rollup/plugin-commonjs": "^21.0.0",
+    "@rollup/plugin-typescript": "^8.3.0",
     "@storybook/addon-actions": "^6.2.7",
     "@storybook/addon-essentials": "^6.2.7",
     "@storybook/addon-links": "^6.2.7",
@@ -56,15 +58,17 @@
     "json-schema-to-typescript": "^9.1.1",
     "lint-staged": "^10.2.11",
     "mini-css-extract-plugin": "^1.3.8",
+    "postcss": "^8.3.9",
     "prettier": "^2.3.1",
     "pretty-quick": "^2.0.1",
     "react-docgen-typescript-loader": "^3.6.0",
     "react-hot-loader": "^4.12.18",
     "react-is": "^17.0.1",
+    "rollup-plugin-postcss": "^4.0.1",
     "sloc": "^0.2.1",
     "ts-loader": "^6.2.1",
     "type-coverage": "^2.17.2",
-    "typescript": "4.4.2",
+    "typescript": "4.4.3",
     "typescript-plugin-css-modules": "^3.2.0",
     "utility-types": "^3.10.0",
     "vite": "^2.5.0",
@@ -76,7 +80,11 @@
   },
   "scripts": {
     "codechecks": "codechecks",
-    "build-modules": "tsc -b -v tsconfig.json",
+    "build": "yarn webpack-build",
+    "build-types": "node tools/build-types.js",
+    "build-storybook": "build-storybook --no-dll",
+    "build-rollup": "rollup -c rollup.config.js",
+    "build-tsc": "tsc -b -v tsconfig.json",
     "type-coverage": "type-coverage",
     "webpack-build": "NODE_ENV=production webpack --mode production",
     "watch": "webpack -w",
@@ -88,9 +96,6 @@
     "lint": "eslint 'packages/**/*.{ts,tsx}' 'demo/src/*.{ts,tsx}'",
     "start-storybook": "start-storybook -p 6006",
     "clean": "rm -rf packages/*/out .buildcache",
-    "build": "yarn webpack-build",
-    "build-types": "node tools/build-types.js",
-    "build-storybook": "build-storybook --no-dll",
     "storybook": "start-storybook -p 6006 --no-dll",
     "sloc": "sloc packages demo tools -e out"
   },
