+++ packages/core/tsconfig.rollup.json
@@ -2,8 +2,8 @@
   "extends": "../tsconfig-base.json",
   "exclude": ["dist", "out"],
   "compilerOptions": {
-    "outDir": "out",
+    "outDir": "dist",
     "rootDir": "src",
-    "tsBuildInfoFile": "../../.buildcache/core.tsbuildinfo"
+    "declarationDir": ".",
   },
 }
