+++ packages/editor/tsconfig.rollup.json
@@ -2,9 +2,9 @@
   "extends": "../tsconfig-base.json",
   "exclude": ["dist", "out"],
   "compilerOptions": {
-    "outDir": "out",
+    "outDir": "dist",
     "rootDir": "src",
-    "tsBuildInfoFile": "../../.buildcache/editor.tsbuildinfo"
+    "declarationDir": ".",
   },
   "references": [
     {"path": "../core"},
