+++ packages/parser/src/global.d.ts
