+++ packages/react/src/global.d.ts
