+++ packages/semantic/src/global.d.ts
