+++ packages/solver/src/global.d.ts
