+++ packages/testing/tsconfig.json
@@ -1,8 +1,9 @@
 {
   "extends": "../tsconfig-base.json",
+  "exclude": ["out", "dist"],
   "compilerOptions": {
-    "outDir": "./out",
-    "rootDir": "./src",
+    "outDir": "out",
+    "rootDir": "src",
     "tsBuildInfoFile": "../../.buildcache/testing.tsbuildinfo"
   },
   "references": [
