+++ packages/testing/tsconfig.rollup.json
@@ -1,13 +1,14 @@
 {
   "extends": "../tsconfig-base.json",
-  "exclude": ["dist", "out"],
+  "exclude": ["out", "dist"],
   "compilerOptions": {
-    "outDir": "out",
+    "outDir": "dist",
     "rootDir": "src",
-    "tsBuildInfoFile": "../../.buildcache/parser.tsbuildinfo"
+    "declarationDir": ".",
   },
   "references": [
     {"path": "../core"},
+    {"path": "../parser"},
     {"path": "../semantic"}
   ]
 }
