+++ packages/typesetter/src/typesetters/delimited.ts
@@ -2,15 +2,14 @@ import * as Editor from "@math-blocks/editor";
 import type {Mutable} from "utility-types";
 
 import * as Layout from "../layout";
-import {makeDelimiter} from "../utils";
 
-import type {Context} from "../types";
+import type {Context, HBox, Glyph} from "../types";
 
 export const typesetDelimited = (
-    typesetChild: (index: number, context: Context) => Layout.HBox | null,
+    typesetChild: (index: number, context: Context) => HBox | null,
     node: Editor.types.CharDelimited | Editor.ZDelimited,
     context: Context,
-): Layout.HBox => {
+): HBox => {
     const thresholdOptions = {
         value: "both" as const,
         strict: true,
@@ -21,19 +20,19 @@ export const typesetDelimited = (
         throw new Error("Delimited's content should be defined");
     }
 
-    const open = makeDelimiter(
+    const open = Layout.makeDelimiter(
         node.leftDelim.value,
         row,
         thresholdOptions,
         context,
-    ) as Mutable<Layout.Glyph>;
+    ) as Mutable<Glyph>;
 
-    const close = makeDelimiter(
+    const close = Layout.makeDelimiter(
         node.rightDelim.value,
         row,
         thresholdOptions,
         context,
-    ) as Mutable<Layout.Glyph>;
+    ) as Mutable<Glyph>;
 
     open.pending = node.leftDelim.pending;
     close.pending = node.rightDelim.pending;
@@ -41,7 +40,7 @@ export const typesetDelimited = (
     const delimited = Layout.makeStaticHBox(
         [open, row, close],
         context,
-    ) as Mutable<Layout.HBox>;
+    ) as Mutable<HBox>;
 
     delimited.id = node.id;
     delimited.style = node.style;
