+++ vite.config.js
@@ -1,14 +1,14 @@
 import path from "path";
-import { defineConfig } from 'vite'
-import reactRefresh from '@vitejs/plugin-react-refresh'
+import {defineConfig} from "vite";
+import reactRefresh from "@vitejs/plugin-react-refresh";
 
 export default defineConfig({
     plugins: [reactRefresh()],
     resolve: {
         alias: [
-            { 
+            {
                 find: /^@math-blocks\/(.*)$/,
-                replacement: path.join(__dirname, 'packages/$1/src/index.ts'),
+                replacement: path.join(__dirname, "packages/$1/src/index.ts"),
             },
         ],
     },
